Refactor serviceClient logic to clients component
